Fix int to size_t for active_space_size in MP2 container
_compute_one_rdm() and _compute_t2_amplitudes() declared
active_space_size as int. The stride expressions
active_space_size * active_space_size * active_space_size overflow
int32 before being assigned to size_t when norbs > ~1290.
Changed to size_t.

+++ cpp/src/qdk/chemistry/data/wavefunction_containers/mp2.cpp
@@ -114,7 +114,7 @@ void MP2Container::_compute_t1_amplitudes() const {
   auto [n_alpha, n_beta] = get_active_num_electrons();
 
   // Get active space size
-  int active_space_size = get_orbitals()->get_num_molecular_orbitals();
+  size_t active_space_size = get_orbitals()->get_num_molecular_orbitals();
   if (get_orbitals()->has_active_space()) {
     const auto& [active_space_ind_alpha, active_space_ind_beta] =
         get_orbitals()->get_active_space_indices();
@@ -152,7 +152,7 @@ void MP2Container::_compute_t2_amplitudes() const {
   }
 
   // Get active space size
-  int active_space_size = get_orbitals()->get_num_molecular_orbitals();
+  size_t active_space_size = get_orbitals()->get_num_molecular_orbitals();
   if (get_orbitals()->has_active_space()) {
     const auto& [active_space_ind_alpha, active_space_ind_beta] =
         get_orbitals()->get_active_space_indices();
